Git: allow having Markdown #Headers in commit msg
## Context

It's convenient to use markdown in commit messages since it's quite readable markup even in plain text.
Also it's works well for Github PRs as the message gets nicely formatted.

## Problem

By default, git considers messages starting with `#` as a comment to be discarded from the commit text.

As result:
1. missing commit sentences
2. poor formatting in rendered view
3. working around the git configuration

despite some workarounds:
1. pre-pend space ` #` or
2. use other markup

which is error prone.

And as a conclusion:
> using `#` as a comment is a poor default.

## Solution
configure Git to use other comment char:
```
git  config --global core.commentChar '?'
```

## Risks
low: `?` is rarely used as a first character.